% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.r
\name{get_option}
\alias{get_option}
\title{Get Package Option Or Default Value}
\usage{
get_option(option)
}
\arguments{
\item{option}{A character string representing the name of the option to
retrieve. See the public fields of \code{\link[R6:R6Class]{R6::R6}} class \code{\link{Options}} for
the list of available \code{\link{parabar}} \code{\link[base:options]{options}}.}
}
\value{
The value of the requested \code{\link[=Options]{option}} present in the
\code{\link[base:options]{base::.Options}} list, or the corresponding default value (i.e., see
\code{\link{Options}}). If the requested \code{\link[=Options]{option}} is not
known, an error is thrown.
}
\description{
This function is a helper for retrieving the value of \code{\link{parabar}}
\code{\link[base:options]{options}}. If the \code{\link[=Options]{option}} requested
is not available in the session \code{\link[base:options]{base::.Options}} list, the corresponding
default value set by the \code{\link{Options}} \code{\link[R6:R6Class]{R6::R6}} class is returned
instead.
}
\seealso{
\code{\link{Options}}, \code{\link[=set_default_options]{set_default_options()}}, \code{\link[base:options]{base::options()}},
\code{\link[=base:getOptions]{base:getOptions()}}
}
