% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{Helper}
\alias{Helper}
\title{Package Helpers}
\format{
\describe{
\item{\code{Helper$get_class_name()}}{Helper for getting the class of a given object.}
\item{\code{Helper$get_option()}}{Get package option, or corresponding default value.}
\item{\code{Helper$set_option()}}{Set package option.}
\item{\code{Helper$check_object_type()}}{Check the type of a given object.}
}
}
\description{
This class contains static helper methods.
}

