% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\docType{data}
\name{LOGO}
\alias{LOGO}
\title{The Package Logo}
\format{
An object of class \code{character} containing the \code{ASCII} logo.
}
\usage{
LOGO
}
\description{
The logo is generated by \code{\link[=make_logo]{make_logo()}} and displayed on package
attach for interactive \code{R} sessions.
}
\examples{
print(LOGO)

}
\seealso{
\code{\link[=make_logo]{make_logo()}}
}
\keyword{datasets}
