% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{make_logo}
\alias{make_logo}
\title{Generate Package Logo}
\usage{
make_logo(
  template = "./inst/assets/logo/parabar-logo.txt",
  version = c(1, 0, 0)
)
}
\arguments{
\item{template}{A character string representing the path to the logo
template.}

\item{version}{A numerical vector of three positive integers representing the
version of the package to append to the logo.}
}
\value{
The ASCII logo.
}
\description{
This function is meant for generating or updating the logo. After running
this procedure we end up with what is stored in the \code{\link{LOGO}}
constant.
}
\examples{
\dontrun{

# Generate the logo.
logo <- make_logo()

# Print the logo.
cat(logo)
}

}
\seealso{
\code{\link{LOGO}}
}
