% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.r
\name{set_default_options}
\alias{set_default_options}
\title{Set Default Options}
\usage{
set_default_options()
}
\value{
The function returns void. The \code{\link[base:options]{options}} set can be
consulted via the \code{\link[base:options]{base::.Options}} list. See the \code{\link{Options}}
\code{\link[R6:R6Class]{R6::R6}} class for more information on the default values set by this
function.
}
\description{
This function is used to set the default \code{\link[base:options]{options}} values
for the \code{\link{parabar}} package. The function is automatically called at
package load and the entry created can be retrieved via
\code{\link[base:options]{getOption("parabar")}}. Specific package
\code{\link[base:options]{options}} can be retrieved using the helper function
\code{\link[=get_option]{get_option()}}.
}
\seealso{
\code{\link{Options}}, \code{\link[=get_option]{get_option()}}
}
