% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.r
\name{set_option}
\alias{set_option}
\title{Set Package Option}
\usage{
set_option(option, value)
}
\arguments{
\item{option}{A character string representing the name of the option to
adjust. See the public fields of \code{\link[R6:R6Class]{R6::R6}} class \code{\link{Options}} for
the list of available \code{\link{parabar}} \code{\link[base:options]{options}}.}

\item{value}{The value to set the \code{\link[=Options]{option}} to.}
}
\value{
The function returns void. It throws an error if the requested
\code{\link[=Options]{option}} to be adjusted is not known.
}
\description{
This function is a helper for setting \code{\link{parabar}}
\code{\link[base:options]{options}}. The function adjusts the fields of the
\code{\link{Options}} instance stored in the \code{\link[base:options]{base::.Options}} list. If no
\code{\link{Options}} instance is present in the \code{\link[base:options]{base::.Options}} list, a
new one is created.
}
\examples{
# Get the status of progress tracking.
get_option("progress_track")

# Set the status of progress tracking to `FALSE`.
set_option("progress_track", FALSE)

# Get the status of progress tracking again.
get_option("progress_track")

}
\seealso{
\code{\link{Options}}, \code{\link[=get_option]{get_option()}},
\code{\link[=set_default_options]{set_default_options()}}
}
