% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.r
\name{stop_backend}
\alias{stop_backend}
\title{Stop a Backend}
\usage{
stop_backend(backend)
}
\arguments{
\item{backend}{An object of class \code{\link{Backend}} as returned by the
\code{\link[=start_backend]{start_backend()}} function.}
}
\value{
The function returns void. It throws an error if the
\code{\link[=Backend]{backend}} provided is already stopped.
}
\description{
This function can be used to stop a \code{\link[=Backend]{backend}} created
by \code{\link[=start_backend]{start_backend()}}.
}
\details{
This function is a convenience wrapper around the lower-lever API of
\code{\link{parabar}} aimed at developers. More specifically, this function
calls the \code{stop} method on the provided \code{\link[=Backend]{backend}}
instance.
}
\examples{
# Create a synchronous backend.
backend <- start_backend(cores = 2, cluster_type = "psock", backend_type = "sync")

# Check that the backend is active.
backend$active

# Stop the backend.
stop_backend(backend)

# Check that the backend is not active.
backend$active

}
\seealso{
\code{\link[=start_backend]{start_backend()}}, \code{\link{Service}}
}
