% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Warning.R
\name{Warning}
\alias{Warning}
\title{Package Warnings}
\format{
\describe{
\item{\code{Warning$requested_cluster_cores_too_low()}}{Warning for not requesting enough cluster cores.}
\item{\code{Warning$requested_cluster_cores_too_high()}}{Warning for requesting too many cluster cores.}
\item{\code{Warning$requested_cluster_type_not_supported()}}{Warning for requesting an unsupported cluster type.}
\item{\code{Warning$progress_not_supported_for_backend()}}{Warning for using a backend incompatible with progress tracking.}
}
}
\description{
This class contains static methods for throwing warnings with informative
messages.
}

